Moved large code parts to new files, cleaned code and introduced object database
The objects to be spawned are now spawned from within lisp (and not from the launch file anymore). Service calls to gazebo spawning functions are added. There is now a list *object-list* acting as an object database. In this database, names, types, handles, urdf-file-paths and intial object poses are stored.