Changed the scenario interface to designator and cleaned up the code
The scenario (`pick-and-place-scenario') is now started using an object-designator rather than the name of the object (string). This was done in favour of getting rid of the function `perceive-named-object' which was only implemented for interface-reasons (and adds no computational value at all). The scenario can now be started using the function `start-scenario' which takes the object-name as it's only parameter and generates the necessary object-designator to start the actual scenario plan. The overall code was cleaned and restructured a bit.