Created two methods for `start-scenario' for working with names or types
When starting the scenario, either an object name can be given as a parameter or an object type. When a name is given, the explicit object instance with that name will be grasped. When a type is given, the first object from the knowledge base with that type will be grasped.