Extended the scenario to try multiple instances of an object in question
When supplied with a valid object description via a designator, the scenario function now goes through all instances that were found to match the description. Each instance is tried to be grasped. Should one succeed, it is taken and the scenario continues. Should this fail, the next instance is tried. The scenario ultimately fails is no instance could be grasped. As named instances (e.g. `mug1') are unique, these objects are specifically only tried once and the scenario fails when grasping doesn't succeed. Instances referenced by their type could very well be numerous. Therefore, this case is handled as `give me any one instance of this type'. All instances are tried in order until one succeeds or all fail.