Added a (occupancy grid) map with a bounding rectangle around the scene
The former map (map.png) just included the two tables included in the scenario. Since we want the robot to operate on only one side of the tables, a thick bounding rectangle is now present around the valid freespace scenery, forcing the robot to stay inside. This way, location costmaps are not generating valid base positions outside of the scene.